% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedEquations.R
\name{secondSensitivities}
\alias{secondSensitivities}
\title{Compute the second sensitivities of the ODEs}
\usage{
secondSensitivities(sensParVariablesY0, sensParVariablesP, dv, dp, ds,
  variables, pars, states, Dy0y, Dpy, Dyf, newfun, is.zero.sens)
}
\value{
the second sensitivities as names character vector with attribute "secondOutputs"
}
\description{
Compute the second sensitivities of the ODEs
}
\details{
Compute second sensitivity equations with respect to parameters and initial values
}

